telegram_bot/requirements.txt
python-telegram-bot>=20.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
